Type API integration helpers
